fix(geometry): Use start point of conic to initialize parameter (#98)
In the ChopIntoQuadsPOW2 method, a given conic is split into two quads,
which produces five points from three. The first point is not used and
should be a start point, while the remaining four points represent the
two quads. At the end of the method, validation should only apply to the
last four points. If we validate all five points under the current
logic, the first point must be initialized as a start point.