chore(wasm): Remove wasm binding code and link config
Remove the legacy wasm binding code and associate link config.
These bindings were used to debug the rendering in webgl environment and have not been maintained for a long time.
Moreover, they are poorly designed.

The WebGL and WebGPU binding APIs for use in the browser environment will be provided in a separate repository.